tests configuration

+++ src/main/java/epam/application/FacadeGymCrmSystem.java
@@ -3,19 +3,26 @@ package epam.application;
 import epam.domain.Trainee;
 import epam.domain.Trainer;
 import epam.domain.Training;
+import epam.domain.TrainingType;
+import epam.exception.UnauthorizedException;
 import epam.request.TraineeDTO;
 import epam.request.TrainerDTO;
 import epam.request.TrainingDTO;
+import epam.request.TrainingTypeDTO;
 import epam.service.TraineeService;
 import epam.service.TrainerService;
 import epam.service.TrainingService;
+import epam.service.TrainingTypeService;
 import epam.util.TraineeMapper;
 import epam.util.TrainerMapper;
 import epam.util.TrainingMapper;
+import epam.util.TrainingTypeMapper;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Logger;
 
@@ -29,18 +36,25 @@ public class FacadeGymCrmSystem {
     private final TraineeMapper traineeMapper;
     private final TrainerMapper trainerMapper;
     private final TrainingMapper trainingMapper;
+    private final TrainingTypeMapper trainingTypeMapper;
+    private final TrainingTypeService trainingTypeService;
 
     private static final Logger logger = Logger.getLogger(FacadeGymCrmSystem.class.getName());
 
+    @Autowired
     public FacadeGymCrmSystem(TrainerService trainerService, TraineeService traineeService,
                               TrainingService trainingService, TraineeMapper traineeMapper,
-                              TrainerMapper trainerMapper, TrainingMapper trainingMapper) {
+                              TrainerMapper trainerMapper, TrainingMapper trainingMapper,
+                              TrainingTypeMapper trainingTypeMapper,
+                              TrainingTypeService trainingTypeService) {
         this.trainerService = trainerService;
         this.traineeService = traineeService;
         this.trainingService = trainingService;
         this.traineeMapper = traineeMapper;
         this.trainerMapper = trainerMapper;
         this.trainingMapper = trainingMapper;
+        this.trainingTypeMapper = trainingTypeMapper;
+        this.trainingTypeService = trainingTypeService;
     }
 
     public Trainee createTrainee(TraineeDTO traineeDTO) {
@@ -54,24 +68,74 @@ public class FacadeGymCrmSystem {
         return traineeService.authenticateTrainee(username, password);
     }
 
-    public Trainee updateTrainee(TraineeDTO traineeDTO, Long userId) {
-        log.info("Updating trainee with id {}", userId);
+    public Trainee updateTrainee(TraineeDTO traineeDTO, String username, String password, Long userId) {
+        log.info("Updating trainee with username {} {}", username, password);
+        if (!authenticateTrainee(username, password)) {
+            log.warn("Unauthorized access denied for user name {}", username);
+            throw new UnauthorizedException("User is not authenticated");
+        }
         var trainee = traineeMapper.toModel(traineeDTO);
+
         return traineeService.updateProfile(trainee, userId);
     }
 
-    public Trainee getTrainee(Long userId) {
-        logger.info("Getting trainee with userId " + userId);
+    public Trainee getTrainee(String username, String password, Long userId) {
+        log.info("Getting trainee with userId {}", userId);
+        if (!authenticateTrainee(username, password)) {
+            log.warn("Unauthorized access denied for user name {}", username);
+            throw new UnauthorizedException("User is not authenticated");
+        }
         return traineeService.findById(userId);
     }
 
-    public void deleteTrainee(String username) {
+    public void deleteTrainee(String username, String password) {
         log.info("Deleting trainee with user name {}", username);
+        if (!authenticateTrainee(username, password)) {
+            log.warn("Unauthorized access denied for user name {}", username);
+            throw new UnauthorizedException("User is not authenticated");
+        }
         traineeService.deleteProfile(username);
     }
 
+    public void activateTrainee(String username, String password, Long id) {
+        log.info("Activating trainee with user name {}", username);
+        if (!authenticateTrainee(username, password)) {
+            log.warn("Unauthorized access denied for user name {}", username);
+            throw new UnauthorizedException("User is not authenticated");
+        }
+        traineeService.activate(id);
+    }
+
+    public void deactivateTrainee(String username, String password, Long id) {
+        log.info("Deactivating trainee with user name {}", username);
+        if (!authenticateTrainee(username, password)) {
+            log.warn("Unauthorized access denied for user name {}", username);
+            throw new UnauthorizedException("User is not authenticated");
+        }
+        traineeService.deactivate(id);
+    }
+
+    public void changeTraineePassword(String username, String password, String newPassword) {
+        log.info("Changing password trainee with user name {}", username);
+        if (!authenticateTrainee(username, password)) {
+            log.warn("Unauthorized access denied for user name {}", username);
+            throw new UnauthorizedException("User is not authenticated");
+        }
+        traineeService.changePassword(username, newPassword);
+    }
+
+    public void createTrainingType(List<TrainingTypeDTO> trainingTypeDTOs) {
+        log.info("Creating training types ");
+        List<TrainingType> trainingTypeList = new ArrayList<>();
+        for (TrainingTypeDTO trainingTypeDTO: trainingTypeDTOs) {
+            var trainingType = trainingTypeMapper.toModel(trainingTypeDTO);
+            trainingTypeList.add(trainingType);
+        }
+        trainingTypeService.saveTrainingType(trainingTypeList);
+    }
+
     public Trainer createTrainer(TrainerDTO trainerDto) {
-        logger.info("Creating trainer " + trainerDto.getFirstName() + " " + trainerDto.getLastName());
+        log.info("Creating trainer {} {}", trainerDto.getFirstName(), trainerDto.getLastName());
         var trainer =  trainerMapper.toModel(trainerDto);
         return trainerService.save(trainer);
     }
